# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a backend server for Digital Achievement Ledger.",
        "title": "Digital Achievement Ledger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/achievements": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of achievements",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "achievements"
                ],
                "summary": "Get all achievements",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by student ID",
                        "name": "student_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by competition name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by category ID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by level ID",
                        "name": "level_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by year",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.PaginatedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new achievement record for a student",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "achievements"
                ],
                "summary": "Create an achievement",
                "parameters": [
                    {
                        "description": "Achievement creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_model.CreateAchievementRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    }
                }
            }
        },
        "/achievements/attachments/{attachmentId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove an attachment from an achievement",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "achievements"
                ],
                "summary": "Delete achievement attachment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Attachment ID",
                        "name": "attachmentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    }
                }
            }
        },
        "/achievements/categories": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of available achievement categories (e.g., Akademik, Olahraga)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "achievements"
                ],
                "summary": "Get achievement categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    }
                }
            }
        },
        "/achievements/levels": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of available achievement levels (e.g., Nasional, Provinsi)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "achievements"
                ],
                "summary": "Get achievement levels",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    }
                }
            }
        },
        "/achievements/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed info about a specific achievement",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "achievements"
                ],
                "summary": "Get achievement by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Achievement ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update details of an existing achievement",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "achievements"
                ],
                "summary": "Update an achievement",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Achievement ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Achievement update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_model.UpdateAchievementRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a specific achievement and all its attachments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "achievements"
                ],
                "summary": "Delete an achievement",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Achievement ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    }
                }
            }
        },
        "/achievements/{id}/attachments": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload an attachment (image/pdf) for an achievement",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "achievements"
                ],
                "summary": "Upload achievement attachment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Achievement ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Attachment file (PDF, JPG, PNG)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Optional label/description for the attachment",
                        "name": "label",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_ahmadqo_digital-achievement-ledger_internal_model.CreateAchievementRequest": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "competition_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "level_id": {
                    "type": "integer"
                },
                "organizer": {
                    "type": "string"
                },
                "rank": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_ahmadqo_digital-achievement-ledger_internal_model.UpdateAchievementRequest": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "competition_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "level_id": {
                    "type": "integer"
                },
                "organizer": {
                    "type": "string"
                },
                "rank": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_ahmadqo_digital-achievement-ledger_internal_response.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Pagination"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_ahmadqo_digital-achievement-ledger_internal_response.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_ahmadqo_digital-achievement-ledger_internal_response.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "errors": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}